BT Image Fix
Fixes an issue where two transparent black CLUT values were used in the sprites for techniques, resulting in oddities in game. This fix is a bit hacky, and will likely result in a 1.0.3 to clean up the issue later and make it a togglable value. Right now, checks for "bt" in the filename to apply the fix.