Battle Texture CLUT Fix
The tool was filtering out the additional colors used in CLUTS 2, 3, and 4 for battle techniques.
This replaces those values once the rebuild of that TIM is complete, preserving technique backgrounds and INT based attack colors.
This bug fixes an issue only seen when a More Monster monster is in the 2P position during combat.